yummi-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
redis==5.0.8
slowapi==0.1.9
prometheus-fastapi-instrumentator==6.1.0
structlog==24.1.0
openai==1.51.0
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.2
cryptography==43.0.1
sentry-sdk[fastapi]==2.17.0
